+++ Module3/M3U2.html
@@ -6,7 +6,7 @@
 -->
 <html>
 	<head>
-		<title>Unit 1</title>
+		<title>Unit 2</title>
 		<meta charset="utf-8" />
 		<meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no" />
 		<link rel="stylesheet" href="assets/css/main.css" />
