+++ Module3/M3U6.html
@@ -50,7 +50,7 @@
 							<section id="one">
 								<div class="inner">
 									<header class="major">
-										<h1>Unit 1: </h1>
+										<h1>Unit 6: </h1>
 									</header>
 									<span class="image main"><img src="images/pic11.jpg" alt="" /></span>
 									<header class="major">
