+++ index.html
@@ -96,7 +96,7 @@
 									</span>
 									<header class="major">
 										<h3><a href="Aboutme.html" class="link">About me</a></h3>
-										<p>Nisl sed aliquam</p>
+										<p></p>
 									</header>
 								</article>
